Limit tests to mac operator

+++ .ci/code_generation.py
@@ -23,6 +23,8 @@ for file_path in files:
         continue
     if file_path.stem == "translator":
         continue
+    if file_path.stem != "mac":
+        continue
     try:
         if file_path.is_dir():
             shutil.rmtree(file_path)
